Rename and disable crawl optimization route
